worked on fixing text upload with console logging
