+++ api/src/main/java/com/exelaration/abstractmemery/AbstractMemeryApplication.java
@@ -2,8 +2,9 @@ package com.exelaration.abstractmemery;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = {SecurityAutoConfiguration.class})
 public class AbstractMemeryApplication {
 
 	public static void main(String[] args) {
